Re-arrange src folder hierarchy
The name {src,mockup,tests}/java/ was a hold over from long ago,
probly because Eclipse has some definite ideas about what is in a
source folder and what not. Nevertheless, it's not the most useful
name.

With the advent of an Java code generator, move the actual language
bindings code from src/java/ to src/bindings/, and rename the other
directories accordinlgy. 

Update Makefile rules and Eclipse configuration.