Update Makefile to build generator classes
Add stanzas to the top level Makefile to build the newly introduced
code generator classes. These are sent to tmp/generator/ to keep them
isolated from the code that ships, as built to what is now tmp/bindings/
Update .classpath to follow.

Fix path that the top level configure probes to find Version.java

Rename unit test to ValidateUtilityMethods and add to the overall
suite as invoked by [default].UnitTests.