Introduce Block class hierarchy
Create a class hierachy of Blocks, representing each (define-...)
stanza in the .defs file as a Java object.

These classes have protected fields exactly corresponding to the field
names in the s-expressions; translation of the usage of these fields
happends when the appropriate Generators is instantiated and returned
by that Block's generator() method.