Build reflector into Block to process characteristics lines
In order to go from the string key names in the characteristic lines
in a define stanza to assigning that value to a field in a Block
object, Block now has processCharacteristic(), called from the
constructor, which uses reflection to transform the key name into a
setter method lookup and then invokes that setter with the supplied
value.

Add unit tests to ValidateBlockUsage verifiying this behaviour.

Miscallaneous minor fixes to OutputDemo and friends following futher
evolution of the APIs.