Refactor addition of pointer '*' to create{Thing,Generator}()
Instead of adding the pointer '*' symbok in  Block's setOfObject()
(which was a messy place for it), do it instead in a utility method
Block.addPointerSymbol() and call it from the createThing() and
createGenerator() methods in the various Block subclasses instead.

This makes it clear that the create methods which call the Thing and
Generator constructors is where the mapping happends from data that is
in the .defs file to data in names and form that is as used by us in
the code generator.