Further work on the defs file parser.
Iterative work taking DefsParser closer to the necessary state. Shift
to working across tests/generator/GtkButton.defs, which we can doctor
as necessary until we have the full two pass parser in place.

Add EnumBlock and EnumGenerator, the former not yet having the code to
process the individual values, and the latter being a no-op stub.
Nevertheless, moves from a NullPointerException in DefsParser to
problems particular to (define-enum ...) not being finished.

Add two more primative types to Thing's static {...} block.