Replace Thing's translationCode with abstract methods
Thing had a field called translationCode containing "pointerOf" in the
case of a Proxy, but that wasn't much help for Constant (a more
complicated call) and in any case omitted to deal with the return
direction.

So replace the field with two abstract methods: translationToNative()
and translationToJava(), imlemented appropriately in the concrete
subclasses.