Rename ofObject field to objectType; change constructors
The Generator constructors had ofObject parameters which set
FunctionGenerator.ofObject which was a poor name. In particular, that
is the name in (some) of the .defs Blocks, and not terribly useful for
us when trying to find snippets when writting code output.

Rename the constructor parameter to gObjectType; refactor the field to
Thing objectType, nicely paralleling Thing returnType there as well.