Import code for VirtualGenerator
Add new code generator VirtualGenerator to output the code for the
Java side of how we land signal callbacks. While this subclasses
FunctioGenerator, it pretty much does its own thing when it comes to
its actual output sequence.

Fix up VirtualBlock to instantiate a VirtualGenerator properly.

This code is what motivated the refactoring of the former
translationCode field in Thing; add some additional unit tests to
ValidateUtilityMethods to keep tabs of this work.