Refactor public method names; push ObjectGenerator code up
In order that we can insert the include statements, we need to
generate the file headers seperately from the class bodies. So, split
writeJava() into writeJavaHeader() and writeJavaBody(), and do
likewise with writeC().

Push almost all the code from ObjectGenerator up to TypeGenerator, as
the lead in for the translation and JNI layers for the java-gnome bindings is
almost identical across the board.