Rearrange into codegen and defsparser packages
To help improve the comprehensibility of the code generator, split the
former single codegen package into defsparser and codegen packages.
This way the class tree of Blocks is moved to defsparser and is
isolated from the effectively parallel class tree for Generator,
staying in codegen. The Thing tree (which is ultimately just an
internal data type) is in codegen as well.

Most of the classes (and their constructors) had to bump to public
visibility as a result. That's ok.

In general, the code in tests/ continues to valiate against package
visible methods. Cut up the unit tests somewhat in order that those
tests that need to see into Thing's package visible fields can do so:
new unit test class ValidateThingUsage is the result.