Preliminary support for out-parameters
This is still very much preliminary, but add the beginnings of what to
do when we hit an out-parameter, including new Thing subclass
OutParameterFundamentalThing, along with the code output in
FunctionGenerator to suit.

Refactor the JNI bit that returns on error to a private method in
FunctionGenerator.

Re-expose the gfloat* out parameter case of gtk_button_get_alignment()
in the test .defs file.