Lower the priority of the make that Eclipse fires off
The build-on-save thing is cool, but when you're in the midst of rapid
changes it results in a lot of makes running that you know probably
won't succeed. So prefix make with nice.

We may need to change the targets around a bit so that the code
generator doesn't always run once we switch it from being a no-op
instruction, but for now nicing it down a bit is good citizenship.