Have BindingsGenerator iterate over multiple files
Split the demo test .defs file in tests/generator/ into three pieces,
more closely simulating the real things. 

Alter the driver code in BindingsGenerator to iterate over these
files. This allows changing the logic for two passes to be closer to
what it is going to look like should we land on that design for real.