Fix missing return type variable declaration in JNI code
Add the overlooked return type declaration needed in
FunctionGenerator's jniFunctionConversionCode().

Since our translation of constructors breaks the convention used
elsewhere in the java-gnome architecture, we need to take care of the
return type declaration as a special case in ConstructorGenerator.
This necessitated re-introducing gReturnType to the constructor and
adjusting ConstructorBlock accordingly. We then override
jniFunctionConversionCode() in ConstructorGenerator to adjust
conditions before calling the super implementation.