Refactoring of Block-Generator and Parser-BindingsGenerator relationships.
It changes the way Block's extracted from .defs are sent to 
BindingsGenerator. Instead of a array of blocks, the Parser now split 
the blocks in two type: the TypeBlock, a single one for .defs file, and 
FunctionBlocks, many of them per .defs file. The FunctionBlock's are 
returned in a List, as a member of the TypeBlock they're related to.

The Generator interface is simpler now. It only have 2 methods: on to 
generate C code, another for Java code. The Bindings generator only call 
both methods in TypeBlock's extracted from .defs, never for 
FunctionsBlocks. In fact, BindingsGenerator never knows about 
FunctionBlocks in this new apporach. Of course, a previous register step 
is still needed.

The TypeGenerator takes care of the generation of code for 
FunctionBlocks. It gets the generator (really a FunctionGenerator) for 
each of them, and invoques it in the right place. Of course, this adds a 
new dependance from TypeGenerator to TypeBlock. In my opinion, this is 
not a problem, but an thing we can even take more advantage of in a 
future.

Now all code is generated by Generator's, no more code is written in 
BindingsGenerator. The management of imports is quite better.