Introduce DefsFile wrapper; refactor Generators
As a bridge between the defs parser and code generator parts of the
program, add new class DefsFile (I'm sure we can come up with a better
name). DefsFile is instantiated with a Block[], and makes sure that it
has what it needs to represent all the information for that type.
Along the way we moved the Thing registration step here.

The first argument to all of the Generator constructor methods used to
be a string indicating the type to which that Block pertained. We now
pass that information by simply putting a reference to the DefsFile as
the first argument. This is in turn provided by createGenerator() in
the Blocks.

All the driving logic is moved from top level BindingsGenerator to the
new class DefsFile. By handing a reference to the current DefsFile
object down to the Generators, they have a much better idea of their
context. This allowed us to massively refactor all the writeXXX()
methods, recollapsing them to one single one for Java, and one single
one for C, named writeTranslationCode() and writeJniCode()
respectively.