Add support for flags.
This adds support for (define-flags ... blocks. Flags are like enums, 
but their values can't be automatically predicted and thus we need to 
call to a JNI method for getting it. In public API, flags are to we 
wrapped as a Constant subclasses, just like enums. 

The Thing for a flag is the same as for a Enum, because both are the 
same thing and are wrapped in the same way.