Improve error messages reported by the defs parser
Switch to a LineNumberReader so that we can report better error
messages when something doesn't parse. This involved refactoring the
regex section in DefsParser from parseStanza() up to readNextStanza().

Add DefsParseException which takes a line number and the actual source
line that failed to parse so we can propegate it up. Introduce
DefsLineNumberReader so that we know what file is being acted on.