Add the necessary casting before returning a java object.
Translation methods that return a Object or Constant make a call to 
objectFor() or constantFor() to translate the type returned form JNI 
method (long or int) to the necessary Java Object. But both constantFor 
and objectFor return generic Constant and Object, while the translation 
methods return a specific subclass. Thus, we need to add the specific 
cast to prevent the compile error.

An alternative to this patch is to add the casting as part of 
translationToJava method in Thing. Following suggestion from AfC, I 
finally choose to put the cast in the FunctionGenerator.