First step to support boxeds.
First, it modifies org.gnome.glib.Plumbing in order to support 
retrieving of a Proxy for a Boxed, in addiction to Value and Object 
proxies. Given that both Boxed and Object proxies are getted in the same 
way (first look for a register proxy, if not try to instantiate 
apropriate Java class for corresponding GType), this refactors objectFor 
as two methods: objectFor and boxedFor, that uses a common proxyFor 
method.

In addiction, this adds code generation stuff for GBoxeds. Things for 
both GObjects and GBoxeds are refactoring adding a common parent class, 
ProxiedThing. Access to fields is still missing.