Add stuff for code generation for boxeds.
This adds the needed stuff for the generation of both translation and 
JNI code for boxeds. All this code is a very first approach that needs 
lots of improves. Mainly these are related to join together, maybe in 
a new class some of stuff that is like that in FunctionGenerator.