Resolve const- prefix problem
The original Thing database design had us manually registering an
additional Thing for the lookup "const-gchar*" in addtion to "gchar*".
This worked fine until we realized that the const- prefix shows up on
other types as well.

To resolve the problem, teach Thing.lookup() to be create and register
a variant Thing object to go with the const case should it be asked to
look it up.

New method createConstVariant() uses reflection to instantiate an
appropriately fully derived clone of the original Thing object. This
necessitated adding protected empty no-arg constructors to all the
Thing subclasses.