Synchronize remainder of mockup; improve comments
Add synchronized blocks to remainder of mockup translation layer
classes.

Improve comments in Gtk.java, Gtk.c and bindings_java_threads.c to
explain as best as we can the operation of the GTK lock as held on the
Java side. Note in particular remarks about the impact of the implicit
wait() in gtk_main().