Refactor Boxed{Block,Generator} to {Getter,Setter}{Block,Generator}
Rather than duplicating all the work of creating methods at
translation, native, and jni layers, refactor BoxedGenerator to only
be the type information and implement the field get and set by simply
overriding jniFunctionLibraryCall() in new classes GetterGenerator and
SetterGenerator, respectively.

[These names may have to change later, but they'll do for now]

To drive this, instead of storing the fields information in BoxedBlock
as would normally have been expected of a (define-boxed...) block,
have DefsParser instantiate GetterBlock and SetterBlock for each
field. Note that creating setters is just for testing; most GTK struct
fields are read only and the parser (and .defs data) will need to be
improved to deal with this.