Further wading through the src/defs/ data
Unfortunately, there are a considerable number of gtypes that we just
don't know how to handle in the code generator. Most of these are out
parameters of various types, so in all likelihood an approach similar
to that taken in Thing.register() for "const-" prefixes will be needed
for * suffixes on types.

Create TypedefFundamentalThing for types that are created as aliases
around otherwise well known fundamental types.

Create FixmeThing for types we expect to be able to fix, and
BlacklistedThing for cases where we know it's a long shot. In both
cases, we need to place a throws RuntimeException into the translation
layer to abort the call.

Not finished yet. If you want to continue working on this, set the dir
in BindingsGenerator.java to "src/defs/" instead of "tests/defs/", and
note src/util/java-gnome.fixes for types that are missing from the
upstream pygtk data.