Added a test that demonstrates releasing lock on callbacks fails.
The release of Gtk+ lock before calling back to java (for example, on a 
signal callback) was an experimental and interesting idea. However, as 
Owen Taylor has told us, is really not a good idea, as causes Gtk+ to 
break. The problem is that Gtk+ has some container signals implemented 
as an iteration over children, calling back signal handler for each 
child. If lock is released before executing the callback, another Thread 
can remove a child while main thread is inside the callback. When main 
thread returns from the callback, the container structure has changed, 
and the iteration can fail with SIGSEGV. Sure a bad idea.

This commit adds a test, OwenProblem, that exposes this problem.