Remove the experimental realease of lock before callbacks.
Given that releasing the lock leads to serious problems, we finally need 
to forget our idea, and just do what is expected to be done in Gtk: 
execute callbacks holding the lock.

Developers must take care about this fact, as forgetting this can lead 
to deadlocks in some situations with at least one application specific 
lock.