Dynamically register missing types as blacklisted
When a type is encountered in a block that we have no information for,
we used to throw IllegalStateException. This requires total closure on
the type data, though, which is rather Catherdral. So instead, change
Thing.lookup() to create a new BlacklistedThing object (which is now
just a wrapper way of making an otherwise blank Thing with
.blacklisted set to true) on the fly.

Also have Thing.lookup() make an attempt to deal with out parameters
dynamically: if a trailing '*' is detected on an otherwise known type,
treat it as an out-parameter or array (which, hopefully luckily) are
the same thing to us (I expect). The code to implement this will be in
createOutVariant() but is incomplete and so sets blacklisted to true
for the time being.

Add two new classes in org.freedesktop.bindings: FIXME (what a name!)
and BlacklistedMethodError. These are used to mark the generated
methods output from the generator if they contain blacklisted types.

I've switched BindingsGenerator to feed off of src/defs/, but you can
easily switch it back to tests/defs/ if you're concentrating on a
specific aspect of the generated output.