Add code to output public API stubs
Add code to the TypeGenerators to enable them to output public API
class stubs in the unique situation where coverage of a new library is
being initiated.

I have only committed this VERY reluctantly. Do not abuse this code
path to try and autogenerate a "complete" public API layer. By our
definition and architecture, it cannot and must not be done. Don't
try. From the comment in DefsFile: "In other words, if you're not the
maintainer of java-gnome, don't touch. I'm really not kidding. Any
patches changing the visibility of this method or adding to its
cability will be immediately rejected without further discussion."

Note that the code to drive this has NOT been committed to
BindingsGenerator. Again, don't touch. Thanks.