Ignore qualifiers in sub-characteristic lines when parsing
Tighten up the parser's regular expressions so that qualifiers such as
(null-ok) are [for now] ignored, allowing us to successfully parse a
greater amount of data that was otherwise valid.

Fix incorrect .defs data for GtkWindow constructor. It used

  (properties
    '("type" (optional))

which is just whacked: type is NOT optional and it has a gtype,
GtkWindowType as it happens. Must be another python-ism. So change it
to a proper (parameters...) declaration. This will be necessary
elsewhere.