Remove conflicting mockup of GtkObject.
Remove the null mockup of GtkObject we had, as there are methods in
generated GtkObject.java that may be relevent.

Model WidgetFlags as extending ObjectFlags, although this is mostly to
explain why the flags() calls are on GtkObject in the source .defs
data. See comments at head of WidgetFlags if you're interested.