Add data and public class for Bitmap
GdkBitmap is not a Boxed (any more), and as a subclass of GdkDrawable,
it deserves a .defs entry, even though it has no methods of its own.
Remove incorrect entry from Thing's static block.

Add cast from [fake] java.lang.Object to FIXME in BlacklistedThing so
that receiver methods compile.