Upgrade build system to work on generated code
Update the project Makefile to work on the translation and JNI layers
now being output by the code generator.

This is to be considered incomplete; we don't have a full build
running at the moment without compile errors, so I was only able to
test it by forcing certain states.

There is also an interesting question relating to the fact that
generated/bindings/ doesn't have any *.java in it on a fresh run. That
doesn't appear to be a problem as the -sourcepath argument to the
compiler seems to pick it all up anyway. Of course, there always is
the possibility that someday we could adjust it to only jar and link
classes and objects that are actually used by the public API, but
that's for another day.

Update Eclipse .classpath to ensure output goes to the right place.