Fix for the import conflict problem.
When two different objects in different packages share the same class 
name in java, they can't be imported both times in a single java file. 
This was a problem in codegen.

This commit fixes that problem. The DefsFile will keep a Set of Thing 
whose plain java class name causes an import conflict. By accessing that 
Set via appropriate functions, the imports for that types are omitted, 
and each time the type needs to be referenced, the fully qualified java 
name is used instead.

Be aware that this commit only handles this problems in method 
parameters and return values. If some class name conflicts with a type 
used in that class file, it will break. For now, this problem doesn't 
exist, so we can handle it later.