Changes in .defs files to fix code compile problems.
These changes involves:

1. Deletion of .defs for deprecated objects.
2. Addition of a .defs file for GtkTextAppearance, a new C struct that 
have some fields previously in GtkTextAttributes.
3. Change properties for parameters in constructors. Some .defs specify 
the parameters of the constructor functions as a (properties... blocks 
instead of (parameters..., as in most functions. These is not recognized 
by our DefsParse, and also has some problems, because type of each 
"property" is not specified. Thus, we change this, to have parameters 
and not properties.
4. Some other little fixes.