Provide references to self for Boxed accessor methods
Since the pseudo blocks GetterBlock and SetterBlock are not subclasses
of MethodBlock, they didn't have the reference to "self" as the first
parameter that is needed for "self->blah" C code to work.

So, move prependReferenceToSelf() from MethodBlock to FunctionBlock to
make it visible to the Accessor tree, and then mess with the
AcessorBlock and AccessorGenerator subclass constructors so that the
necessary parameter list is ulimately passed along to
FunctionGenerator.

Add "parent" to the constructors of GetterBlock and SetterBlock, as
passed in by DefsParser, so they know what the enclosing ofObject
is, something that is a given in all real blocks.