First step to GList/GSList support.
This adds some initial support for handling GList/GSList types. As a 
first step, I've only implemented support for GList as input parameters 
for Gtk+ function. Moreover, this patch only correcly generates the 
translation layer. JNI code still breaks when GList parameters are used.

The idea begin GList support is the following:

- In public layer and translation methods, a GList is "wrapped" as an 
array of the GList type. Given that Java is type-safe, we need to add 
extra information to .defs files to take into account the type of the 
GList elements. I follow suggestion from Andrew Cowie to represent that 
as GList-XXXX, where XXXX is the type of the elements. This patch only 
modifies that in a single method.

- In native method, the list is passed as an array of longs, containing 
the pointers for each of the objects. The translation code performs the 
conversion.

- In JNI code (still not implemented) the data is copied to a GList.