Refactor import collision code
Chasing an elusive unnecessary fully qualification turned out not to
be a bug, but to figure it out I heavily refactored the import
collision logic. It's ultimately the same algorithm as Vreixo
contributed, but somewhat cleaner and more verbosely named.

As a cosmetic improvement, always ensure the main type for a DefsFile
is imported.