Split Makefile into three pieces
Make doesn't seem to have a means of setting variables after a rule
has run. So in order to pick up the .java and .c files created by the
code generator, splice two fragments off of the toplevel Makefile in
order that they can run the `find` commands later than when the
original Makefile is parsed.