Fix major bug in signal marshaller
It turns out that when signals are registered the reference to self
that we thought we needed to specify by hand is present as the first
argument. As a result we were getting a duplicate entry / off-by-one
error that only showed up if you had a signal handler with more than
one parameter.

The first parameter is reference to self thing seems to be a
convention; hopefully it will be stable as it is actually an
assumption we had to make - notably these parameters are not present
in the .defs data for (define-virtual...) just as they are absent for
(define-method...) blocks.

Note that calling boxedFor() presently results in a segmentation
fault, and this gets hit for DELETE_EVENT. Working on it.