Per-package Plumbing classes; flesh out GdkEvent special case
Add a Plumbing class to each package. This enables a place to put code
handling special cases of instantiation. Have TypeGenerator simply
import a Plumbing class in the same package as the generated class's
package.

Deal with the problem presented by GdkEvent, a union: use the
GdkEventType enum present as the first field in all of the GdkEvent
structures as a discriminator. Get at this using a native call
getEventTypeOrdinal() in new class org.gnome.gdk.Plumbing and use it
to figure out which .class literal to pass to the boxedFor() method in
org.gnome.glib.Plumbing in order to actually create the concrete Event
subclass required.

Add public API stubs representing all of the structures in the
GdkEvent union. Specific .defs data will be needed, of course.

Complete a preliminary version of the public API for EventType; the
underlying documentation is quite sparse, and there is every
possibility that we could restrict the visibility of some or all of
these constants as many seem VERY internal.