Add little unit test for methods returning a GList.
This little unit test validates the conversion of a GList* to a java 
array in methods that return a GList*. Notice that only methods that 
return a GList* of GObjects are tested, GList* of boxeds are still not 
tested. 

Current implementation HAS A IMPORTANT BUG and doesn't pass this test.

Also, this commit adds public coverage for getChildren method in 
org.gnome.gtk.Container.