Fix the GList bug detected with previous test case.
There was a serious bug with functions that return a java array for 
proxied objects. The reason is an extrange (and unknown for a 10 years 
java programmer like me) behavior in the way Java language manages 
arrays. If you have a "class A extends B", you can cast an A array to a 
B array, but not an B array to A, even if you know all contents in the 
array are instances of A. 

Thus, when we cast the Object (or Boxed) array returned by 
objectArrayFor (or boxedArrayFor) Plumbing methods to the subtype array 
returned in a method in translation layer (such as 
GtkContainer.getChildren), it throws a ClassCastException.

The solution is to create the array in the translation methods (where we 
known the subtype), and pass it to the generic methods in Plumbing.

Now the unit test passes.