Change null return to empty array
For method functions that return GLists of things, translate this back
to Java as an empty array rather than propegating it back up as a Java
null. This simplifies the logic of people using such methods as they
don't have to guard against NullPointerException and can just use a
for loop bounded by the array's length.

Update unit test ValidateGListMethods accordingly.

Update and improve the JavaDoc for Container's getChildren() with
changed return condition and further explaination of what to do with a
returned array of Widgets.