Build C code
Add functions to express the logic necessary to build the C side of
java-gnome. The tricky part is, of course, the JNI headers but that
went fairly smoothly. Also add functions to compile C code, and one to
link the shared library.

Refactor filesNeedBuilding() into fileNeedsBuilding(), boolean, wiht
filesNeedBuilding() as a convenience wrapper. Exposing the core
algorithm directly made several rule functions simpler.

Write the output pickle to a tmp file so that if it gets smashed on
interrupt it doesn't force a complete rebuild.