Add public coverage for org.gnome.gdk.WindowState.
This coverage is only intented for hacker reference. Thus, the documentation 
for this class is very poor and can't be considered acceptable.

The idea is to provide an example of what a public API writer can do when 
(s)he found a flag. However, both Andrew Cowie and I think that for many
flags the implementation of or() method might be not needed, as it seems
that it's not very useful OR'ing that flags.