Unit test for validating flag ORing and important bug fix.
This adds an unit test for validating the ORing of flags. Previously 
implemented public API is used for that.

That test shown an important bug, that is fixed by this patch too. The 
problem was that in some of the many refactors it was introduced a 
"throw new UnsupportedOperationException(...)" in 
EnumGenerator.writeJniCode(), because enums don't need a C file. But in 
FlagsGenerator we call super.writeJniCode(), thus throwing the same 
exception. However, we need to output C code for flags.

Finally, some minor changes where made to .defs.