Support "test" and "demo" targets
Add support for extra targets to build the test classes needed
to run the unit tests or the demo applications.

Add a third target "eclipse" to ensure the test code is compiled, but
not to run it as that would be done by a launcher within the context
of an Eclipse session.