Expose alignment methods on Button
Finish exposing setAlignment() and getAlignmentX(), getAlignmentY() on
Button, previously present as a non-functioning stub.

This uses the presently hard coded out-parameter logic from
FunctionGenerator; by putting this in place we have a working unit
test from which to test improving the code generator.

Add ValidateOutParameters as a class to hold related test cases.