Refactor jni conversion code to Thing classes
Preliminary refactoring of the code to covnert from JNI to Glib. This
code has for a while now been embedded in FunctionGenerator. Vreixo
Formoso made the suggestion to move this logic to the Thing tree,
something I resisted  originally mostly because I didn't want full
output code to be there, just snippets.

I seem to have found a naming convention to do this that I am
comfortable with, and while there are one or two kludges in
FunctionGenerator, overall it is much cleaner.

Rename OutParameterFundamentalThing to ArrayFundamentalThing. This is
prepratory to fixing the treatment of arrays and lists in general, but
not necessarily the final name we will end up with.