Expose Window's getScreen(); deal with potential GDK referencing bug
Add initial coverage for Screen ane getting one by exposing Window's
getSreen().

This uncovered the fact that the actual object returned by
gtk_window_get_screen() is of type GdkScreenX11 which we have no
information for. We could  add a .defs file for it, but instead
register it custom, mapping it to its parent, org.gnome.gdk.Screen. Do
this in org.gnome.gdk's Plumbing, which seems again to be the special
case repository.

Screen methods worked fine, but on application termination a horrible
GDK error evenutated. Turning on the memory management debugging, this
appears to be due to GDK itself not taking a reference count to the
GdkScreen; when our Screen goes out of scope the ToggleRef goes to
weak and when the object finalizes the ref count drops to zero,
triggering the GdkScreen to be disposed, and GDK doesn't like that.
Very strange. 

Workaround by manually increasing the ref count on the default screen
at start up in Gtk.gtk_init(). Add test case in new file
org.gnome.gdk.ValidateScreensAndDisplays which should allow us to keep
the issue in mind.