Expose StockId as a wrapper around a gchar* stock-id.
This adds a new public class, StockId, intented to be a wrapper aroung a Gtk+
stock-id, that is just a plain gchar* string. However, instead of dealing 
that as a java.lang.String, creating its own class has several advantages:

1. Improved type-safety
2. Better support for overloading of methods, e.g.: Button(String label) vs
   Button(StockId stockId).
3. Natural placeholder for built-in stock ids.

This also makes some corrections to .defs data and adds a Button(StockId)
constructor.