Improve thread documentation in Gdk; fix pre blocks
Using text from the 4.0.3 NEWS file as a base, document our thread
safety mechanism in org.gnome.gdk.Gdk. In particular, note the caveats
about callbacks and the ability to take the lock yourself if
necessary.

Eclipse versions prior to recent iterations of 3.2.2 have a bug
whereby pre block comments are precessed one character cell to the
right every time they are formatted. I'd forgotten to patch this on my 
new laptop, and so it was necessary to fix a series of <pre> blocks
which had managed to get out of hand.