Workaround Gentoo FEATURES=ccache bug
We are very aggressive about dereferencing symlinks for what are
otherwise and normally good reasons, but in the case of
FEATURES=ccache, this causes /usr/lib/ccache/bin/gcc to be on the
path; dereferencing it returns as /usr/bin/ccache (which IS a
legetimite way of invoking it, but we were hard coded to only use it
as prefix to gcc); of course since that ended up as the CC variable,
things fail.

The temporary workaround is to just use `which gcc` if ccache ends up
as the candidate. The proper fix will be to properly detect this state
and set the CCACHE and CC variables more effectively.