Revert handler API to (handler, source, ...)
A while back we changed the generated handler reciever signature from 

	handleWhatever(Signal handler, long source, ...)

to

	handleWhatever(Signal handlerInstance, long sourceObject, ...)

in order to avoid compile time collisions with signal parameters whose
name were "handler" or "source". The problem with this is that the
name sourceObject propegated up through the interface signatures and
out into the exposed public API and also into the example code. Yuk.

It turns out that there are only about 6 instances in the entire .defs
data where virtuals have parameters whose names collide with "handle"
or "source", so this whole thing was a gross overreaction. 

So, to clean up our public signatures, revert the code in
VirtualGenerator back to the original design and fix the tiny number
of .defs instances where these actually came up. Fix parameter names
in existing public Signal interfaces, and clean up example code.