Better stock item handling in StatusIcon.
This replaces the String methods in StatusIcon with the new Stock class. Note
that we need to think what to do with getStock() method. For now, it just
creates a new Stock with the returned String, but maybe we finally prefer a 
register of stock ids.

Also, this commit adds some new constructors that were missing in .defs.