Add support for null-ok parameters.
This adds the needed support for parameters that can be null, while 
guarding against passing null in a parameters that can't be.

The info is taken from the null-ok in .defs, info that sure needs to be 
reviewed.
Note that GBoxed setters will also need this support, but for no it's 
not implemented.