Adds tests for code generator.
This commit adds some tests that validate the code output by code 
generator. The idea is to have simple .defs with handwritten code, that 
will be checked against the code generated.

That way we have both a place for a template of what we have to output, 
and also a way to validate code generator and ensure no regression after 
adding new features.