Have Stock class maintain registry permitting instance lookups
Add a simple HashMap and add instances of Stock to it as they are
created. This permits new package visible method instanceFor() to be
used elsewhere in GTK allowing these Stock constants to be returned.

This is, obviously, a bit simplistic; if we find ourselves needing
this in other packages or for different classes we may need something
a bit more like the org.freedesktop.bindings.Constant infrastructure.

Remove Stock's equals() overload as it is no longer necessary.

Add a new test case, ValidateStockItems to establish the instance
reference lookups are working.