Change Thing hierarchy and methods.
Current Thing hierarchy was wrong in many aspects. This refactor that to have
a clear and more useful hierarchy.

This also adds some new methods to Thing. The idea is to give more code 
generation responsabilities to Thing, to prevent big ifs and instance of in
Generators.

Note that this commit leaves codegen in a wrong state.