Test for generation of static functions code.
Static functions are functions that are neither methods of an object or
constructors, but that can be related with a given object. Currently, 
this functions are ignored by the parser, and stored in missing.defs.
The idea is that we need to move them to the .defs where the object 
they're related to is defined. This commit, as an example, moves some 
functions related to GtkAccelKey.

Additionally, and following the test-driven methodology, I've added some 
test to later validate the implementation of functions support.