Support for static functions in codegen.
This adds the needed changes to codegen to support static functions. 
All (define-function...) are handled by a single FunctionBlock, 
including constructors. Thus, we remove the ConstructorBlock previously 
used. However, constructors are still generated by its own 
ConstructorGenerator.

Static functions are similar to methods, but we don't need to prepend 
the reference to self.