Remove tests cases comparing generated code to models
While having model files of what the generated output should look like
and comparing the generated output to those models, it is fragile; one
change to the whitespace output by the code generator results in 40
model files that must be changed? I don't think so.

Discussion with testing experts including Robert Collins and Peter
Miller highlighted that the underlying problem is isolation; test
cases evaluating the performance of the desired attribute or factor
in the real libraries at run time will be taken in preference.