Have Thing "extra" methods return default values for ordinary types
Compress the code in the Thing hierarchy by replacing the new
characteristics on things that were exposed as abstract methods to
concrete methods returning the default case which is common for most
classes.

This means that subclasses only need have code for these factors if,
in fact, they really do override the normal behaviour - appropriate
given that these factors are "extra" code.